requirements.txt
pip~=24.0
psycopg2~=2.9.9
python-dotenv~=1.0.1